rm_tuner dev init
